Progress!

+++ J03/weekTwo/PhoneBook/Person.java
@@ -34,7 +34,7 @@ public class Person  {
         return this.phoneNumber;
     }
 
-    public void setPhoneNumber(long phoneNumber) {
+    public void setPhoneNumber(int phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
